Moving GuiTabType enum to gui/enums
